% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{binom_summary}
\alias{binom_summary}
\title{Title}
\usage{
binom_summary(outcome)
}
\arguments{
\item{outcome}{}
}
\description{
Title
}
\examples{


library(dplyr)

df <- data.frame(group = rep(c("a", "b"), each = 20), outcome = c(rbinom(20,1, 0.7), rbinom(20,1,0.2)))

df \%>\%
group_by(group) \%>\%
do(binom_summary(.$outcome))
}
