% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{bootstrapped_summary}
\alias{bootstrapped_summary}
\title{Bootstrapped mean and confidence interval}
\usage{
bootstrapped_summary(data, repeats = 1000)
}
\arguments{
\item{data}{A numeric vector containing the values to be summarized}

\item{repeats}{An integer indicating the number of bootstrap samples to be used}
}
\value{
A data.frame containing a column each for mean, lower 95% confidence interval
 and upper 95% confidence interval.
}
\description{
The function returns the mean and the confidence interval for a numeric vector
}
\examples{

library(dplyr)

test <- data.frame(group = rep(letters[1:2], each = 30),
                               dv = c(rnorm(30, 40, 4), rnorm(30,70,4)))
test \%>\%
   group_by(group) \%>\%
   do(bootstrapped_summary(.$dv))
}
