% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{contr_forward}
\alias{contr_forward}
\title{A contrast matrix for a forward difference coding scheme}
\usage{
contr_forward(k)
}
\arguments{
\item{k}{An integer indicating the number of factor levels.}
}
\value{
A matrix with the coding scheme
}
\description{
The function returns a contrast matrix for a forward difference coding scheme.
}
\examples{

# Returns a contrast matrix for a 4-level factor

contr_forward(4)

}
