% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{model_writer}
\alias{model_writer}
\title{Write a list of models to RDS files}
\usage{
model_writer(model_container, save_path = "./", prefix = "")
}
\arguments{
\item{model_container}{A list of model objects.}

\item{save_path}{File destination for the RDS-files to be saved in}

\item{prefix}{Custom Prefix to the file names.}
}
\value{
Nothing
}
\description{
A function to save models from the list of models fitted with the fit_with()-function.
The function saves one RDS file for each model to the provided path
}
\examples{
\dontrun{
models <- modelr::fit_with(mtcars, lm, "mpg ~ hp")
model_writer(models)
}

}
