% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{multiple_file_reader}
\alias{multiple_file_reader}
\title{Batch read files from a directory}
\usage{
multiple_file_reader(files, extension, reader_function, path = ".", ...)
}
\arguments{
\item{files}{Usually, a vector of character strings with the names of the files to read in}

\item{extension}{A character string for the file extension like csv, tsv. Only csv supported right now.}

\item{reader_function}{The function to use for reading individual files}

\item{path}{A character string indicating the path to the files to read}

\item{...}{Additional parameters for the reader function, like sep or dec}
}
\value{
A data frame (or equivalent structure) with the concatenated files
}
\description{
The function reads a batch of files within one directory and concatenates them into
one data.frame (or equivalent structure). This is useful, when your data of each participant is
stored in a separate file.
}
\examples{

\dontrun{

files_names <- letters[1:3]

# Reads three files with .csv extension in the working directory using readr::read_csv

my_data <- multiple_file_reader(file_names, ".csv", reader::read_csv)


}
}
